+++ Cargo.toml
@@ -14,6 +14,7 @@ aes-gcm = { version = "=0.10.3", default-features = false, features = [
     "getrandom",
     "zeroize",
 ] }
+base64 = "0.22.1"
 chrono = "0.4.41"
 derive_more = { version = "2.0.1", features = ["full"] }
 dirs-next = "2.0.0"
