+++ crates/core/src/models/pdf.rs
@@ -4,6 +4,12 @@ use crate::prelude::*;
 #[derive(Clone, Debug, From, AsRef, PartialEq, Eq, Hash)]
 pub struct Pdf(pub Vec<u8>);
 
+impl HasSample for Pdf {
+    fn sample() -> Self {
+        Self(vec![0xde, 0xad, 0xbe, 0xef]) // Sample PDF data
+    }
+}
+
 #[derive(Clone, Debug, PartialEq, Eq, Hash, TypedBuilder, Getters)]
 pub struct NamedPdf {
     #[builder(setter(into))]
@@ -24,7 +30,7 @@ impl HasSample for NamedPdf {
     fn sample() -> Self {
         Self::builder()
             .prepared_data(PreparedData::sample())
-            .pdf(Pdf(vec![0; 100])) // Sample PDF data
+            .pdf(Pdf::sample()) // Sample PDF data
             .saved_at(PathBuf::from("/tmp/sample_invoice.pdf"))
             .name("sample_invoice.pdf".to_string())
             .build()
